refactor(step2):  WinningLotto 삭제
Lotto로 충분히 대체 가능
